refactor: move buildGroupIdParam to separate file
